Separate all redux action types
